Color sensor working

+++ ColorData.h
@@ -2,8 +2,8 @@
 #include <cstdint>
 
 struct ColorData {
-  uint8_t r, g, b;
+  uint16_t r, g, b;
 
-  ColorData(uint8_t r_, uint8_t g_, uint8_t b_) : r(r_), g(g_), b(b_) {}
+  ColorData(uint16_t r_, uint16_t g_, uint16_t b_) : r(r_), g(g_), b(b_) {}
   ColorData() {}
 };
